backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
motor==3.1.2
python-dotenv==1.0.0
pydantic==1.10.7
pillow==9.5.0